requirements.txt
agent-framework[ollama,openai]
python-dotenv
